MLIBZ-3089, refactored DataStoreMultiInsertTest class, moved common code to base class
